requirements.txt
Django==1.11.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
django-suit==0.2.25
PyJWT==1.5.2
pytz==2017.2
bcrypt==3.1.3
sphinx
sphinx_rtd_theme
django-import-export==0.6.0